Pointing issues and castings
git-svn-id: file:///home/jj/hercules.svn/trunk@6975 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hsccmd.c
@@ -2508,7 +2508,7 @@ u_int i;
     }
 
     /* Process options */
-    for (i = 2; i < argc; i++)
+    for (i = 2; (int)i < argc; i++)
     {
         if (strcaseabbrev("lock", argv[i], 1))
             sysblk.lock_mainstor = 1;
@@ -2568,7 +2568,7 @@ u_int i;
     }
 
     /* Process options */
-    for (i = 2; i < argc; i++)
+    for (i = 2; (int)i < argc; i++)
     {
         if (strcaseabbrev("lock", argv[i], 1))
             sysblk.lock_xpndstor = 1;
